Extracted Hanami::Assets::Asset from Hanami::Static
