+++ lib/hanami/static.rb
@@ -1,5 +1,6 @@
 require 'rack/static'
 require 'hanami/assets/compiler'
+require 'hanami/assets/asset'
 
 module Hanami
   class Static < ::Rack::Static
@@ -10,50 +11,8 @@ module Hanami
       @sources = _sources_from_applications
     end
 
-    class Asset
-      PUBLIC_DIRECTORY = Hanami.public_directory.join('**', '*').to_s.freeze
-
-      # @since x.x.x
-      # @api private
-      URL_SEPARATOR = '/'.freeze
-
-      attr_reader :path, :config, :original
-
-      def initialize(sources, path)
-        @path            = path
-        @prefix, @config = sources.find { |p, _| path.start_with?(p) }
-
-        if @prefix && @config
-          @original = @config.sources.find(@path.sub(@prefix, ''))
-        end
-      end
-
-      def precompile?
-        original && config
-      end
-
-      def exist?
-        return true unless original.nil?
-
-        file_path = path.gsub(URL_SEPARATOR, ::File::SEPARATOR)
-        destination = find_asset do |a|
-          a.end_with?(file_path)
-        end
-
-        !destination.nil?
-      end
-
-      private
-
-      def find_asset
-        Dir[PUBLIC_DIRECTORY].find do |asset|
-          yield asset unless ::File.directory?(asset)
-        end
-      end
-    end
-
     def call(env)
-      asset = Asset.new(@sources, env[PATH_INFO])
+      asset = Assets::Asset.new(@sources, env[PATH_INFO])
 
       if serve?(asset)
         precompile(asset)
